refactor: support 패키지 폴더 구조 변경
