+++ H-Mingle/src/main/java/com/hyundai/hmingle/support/oauth/OauthConnector.java
@@ -1,4 +1,4 @@
-package com.hyundai.hmingle.support;
+package com.hyundai.hmingle.support.oauth;
 
 import com.hyundai.hmingle.support.dto.response.GoogleUserResponse;
