+++ H-Mingle/src/main/java/com/hyundai/hmingle/support/oauth/OauthProvider.java
@@ -1,4 +1,4 @@
-package com.hyundai.hmingle.support;
+package com.hyundai.hmingle.support.oauth;
 
 public interface OauthProvider {
